Fix maven build
${project.parent.basedir} is somewhat whidespread in the internets, but
it does not work, at least not in maven 3.3.9 (upstream at the time of this
writing).

Since maven does interpolation of properties after inheritance, inheriting the
${project.basedir} is not an option either. This is static but the best option
atm.

Further log4j is needed in the compile scope, since the project makes use
of it. ( Not just as a transitive dependcy of impala jdbc ).